Rename io package to better fit its purpose
