+++ src/main/java/Mana/storage/LocalDateTimeTypeAdapter.java
@@ -1,4 +1,4 @@
-package mana.io;
+package mana.storage;
 
 import com.google.gson.TypeAdapter;
 import com.google.gson.stream.JsonReader;
@@ -7,7 +7,6 @@ import com.google.gson.stream.JsonWriter;
 
 import java.io.IOException;
 import java.time.LocalDateTime;
-import java.time.ZonedDateTime;
 import java.time.format.DateTimeFormatter;
 
 /**
